Move ownership to emulator team.
This branch ships with the Android Emulator and is not used by any
android build itself. We need yuv to build WebRTC from source.

Currently we are using a prebuilt that is introducing a series of
drawbacks:

- We need to support multiple platforms mac/win/lin (+arm64) variants
- We want to include all the unit tests to reduce risk of failure
- We plan to take changes from stable webrtc.

There should be no risk to android as this branch is ONLY used by the
emulator team.

Bug: 184770746
Change-Id: I1bcca027e7ac6406f9f1c1ea398b9ecfe882aa87